requirements.txt
Pillow==11.1.0
pygame==2.6.1
